init() now sorts the array
this allows for optimizing the code that looks for the next hit.
right now, it iterates through all hits each timestep looking
for any that must be played, which is inefficient. sorting
the array first means we can iterate through the array sequentially
(we only need to look at the next hit in the array every timestep
instead of all of them), which should save some time.

(this commit only adds the sort. optimizing the code will follow in
a future commit)